Moved the HTTP API to thread rather than process
The non-shared memory of the process was causing Cookie Jar listeners to
not get called. By moving to threads, we have a shared address space.
However, in order to do this, we needed to create a shutdown method for
the HTTP server (rather than just killing the process): this turns out
to be rather ugly in Flask!