Lightly hammer the CouchDB server to test for responsiveness
All CouchDB methods were first doing an initial request to check for
responsiveness and validity. If this failed, then the method would fail.
This change makes it keep trying a set number of times, while blocking
and trying not to hammer the DB too hard, until a valid connection is
established or the failure conditions (set in environment variables) are
met... This makes things a teensy bit more stable, insofar as the
CookieJar won't blow up the first time it encounters an unresponsive
database, while not being too hard on the CouchDB server.

(Note that the upsert queue, which also accommodates for an unresponsive
connection, still exists and has no failure conditions; it will keep
trying until entropy exceeds our ability to compute. As such, the
CookieJar will *only* blow up if CouchDB fails...which is approximately
what we want!)

Resolves #30