HTTP API: Fetch and delete by query string, as well as route
This is for the case of identifiers that begin with a slash, in which
case the route would end up as `/cookiejar//my/identifier` and Flask
doesn't seem to want to parse this. Before, there was a fallback that,
when using say `/cookiejar/my/identifier`, would first try as-is then,
if that didn't work, would try again with a prefixed slash. This is
obviously really dangerous, as identifiers can be arbitrary and could
easily mask each other (thus potentially leading to deleting the wrong
one). This commit closes that security hole.