Raise UnresponsiveCouchDB if query iteration fails
When iterating through query results, the generator doesn't apply the
monkey-patched preflight checking and hence, if the DB goes away in the
meantime, it can break. For `_Bert.get_by_identifier` and
`_Bert.queue_length`, we raise `UnresponsiveCouchDB`. The only time this
has been observed is in `Sofabed._batch`, while calling the former.
However, there are a few other query iterations in `BiscuitTin`; the
most dangerous of which (i.e., they *need* to work) are `_Bert.dequeue`
and `_Ernie.get_metadata`. These should be fixed, but I'm not really
sure how, without making the abstraction even more leaky :(

This potentially fixes #47, but not very elegantly...