Reimplementation scaffolding for CouchDB client
n.b., We remove the initial HEAD request, to test for DB availability,
as it's unnecessary (i.e., the actual request will perform the same
function and do useful work, if it's successful). The intention is to
modify pycouchdb.resource.Resource such that every request made is
repeated on failure, with suitable pauses to allow a busy CouchDB server
to potentially (hopefully!) recover. This is the same as before, but it
will affect *every* request made by the CouchDB client.